// src/generated/prisma/models/Jinis.ts
export type JinisWhereInput = {
  AND?: Prisma.JinisWhereInput | Prisma.JinisWhereInput[]
  OR?: Prisma.JinisWhereInput[]
  NOT?: Prisma.JinisWhereInput | Prisma.JinisWhereInput[]
  id?: Prisma.StringFilter<"Jinis"> | string
  slNo?: Prisma.IntFilter<"Jinis"> | number
  name?: Prisma.StringFilter<"Jinis"> | string
  fatherName?: Prisma.StringFilter<"Jinis"> | string
  phoneNo?: Prisma.StringFilter<"Jinis"> | string
  credit?: Prisma.IntFilter<"Jinis"> | number
  type?: Prisma.EnumJinisTypeFilter<"Jinis"> | $Enums.JinisType
  goldWeight?: Prisma.FloatFilter<"Jinis"> | number
  silverWeight?: Prisma.FloatFilter<"Jinis"> | number
  date?: Prisma.DateTimeFilter<"Jinis"> | Date | string
  active?: Prisma.BoolFilter<"Jinis"> | boolean
  settledAt?: Prisma.DateTimeNullableFilter<"Jinis"> | Date | string | null
  createdById?: Prisma.StringFilter<"Jinis"> | string
  createdAt?: Prisma.DateTimeFilter<"Jinis"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Jinis"> | Date | string
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
  items?: Prisma.JinisItemListRelationFilter
  interests?: Prisma.InterestListRelationFilter
}